requirements.txt
google-genai
httpx
beautifulsoup4
fastembed
feedparser
psutil
discord.py
python-dotenv
flask
